Aggiornati metodi CRUD in SpedizioneServiceImpl

+++ server/src/main/java/com/ai/ecommercej4j/service/impl/SpedizioneServiceImpl.java
@@ -17,6 +17,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.ai.ecommercej4j.repository.SpedizioneRepository;
 import com.ai.ecommercej4j.service.SecurityService;
+import java.util.Collections;
 
 @Service
 public class SpedizioneServiceImpl implements SpedizioneService {
@@ -30,17 +31,28 @@ public class SpedizioneServiceImpl implements SpedizioneService {
     public void createSpedizione(SpedizioneCreateDto dto) {
         //Verifica esistenza token
         while (securityService.checkToken(dto.getToken())) {
-            if (spedizioneRepository.findByCodice(dto.getDati().getCodice()) == null) {
-                //Dopo aver verificato l'esistenza,salvo nel database
-                spedizioneRepository.save(dto.getDati());
+            if (spedizioneRepository.findByCodiceContainingIgnoreCase(dto.getDati().getCodice()) != null) {
+                List<Spedizione> ls = spedizioneRepository
+                        .findByCodiceContainingIgnoreCase(dto.getDati().getCodice());
+                //Dopo aver verificato l'esistenza o meno,salvo nel database
+                if (ls.isEmpty()) {
+                    spedizioneRepository.save(dto.getDati());
+                }
             }
         }
     }
 
     @Override
     public SpedizioneSearchResultsDto searchSpedizione(SpedizioneSearchDto dto) {
-        List<Spedizione> ls = spedizioneRepository.findByCodice(dto.getSearchKey());
-        SpedizioneSearchResultsDto dtossr = new SpedizioneSearchResultsDto(ls);
+        SpedizioneSearchResultsDto dtossr = new SpedizioneSearchResultsDto();
+        //Verifico l'esistenza del dto in ingresso e la validità del token
+        if (dto != null && securityService.checkToken(dto.getToken())) {
+
+            List<Spedizione> ls = spedizioneRepository.findByCodiceContainingIgnoreCase(dto.getSearchKey());
+        } else {
+            //Se non esiste restituisco lista vuota
+            dtossr.setResults(Collections.emptyList());
+        }
         return dtossr;
     }
 
@@ -48,8 +60,10 @@ public class SpedizioneServiceImpl implements SpedizioneService {
     public void deleteSpedizione(SpedizioneDeleteDto dto) {
         /* 1-Verifico se esiste già un dto rispetto a quello che ho 
          Elimino spedizione relativa all'ID*/
-        if (securityService.checkToken(dto.getToken())) {
-            spedizioneRepository.deleteById(dto.getIdToDelete());
+        if (dto != null) {
+            if (securityService.checkToken(dto.getToken())) {
+                spedizioneRepository.deleteById(dto.getIdToDelete());
+            }
         }
     }
 
@@ -59,7 +73,7 @@ public class SpedizioneServiceImpl implements SpedizioneService {
 
         // Verifico esistenza dto e codice spedizione non vuoto
         while (dto != null && dto.getDati() != null && dto.getDati().getCodice() != null) {
-            // Verifico vaalidità token
+            // Verifico validità token
             if (securityService.checkToken(dto.getToken())) {
                 // controllo se l'id del dto non è vuoto                
                 if (dto.getDati().getId() != null) {
